Break out indexed BufferReader logic into partial
